Added history, model in /assets
